@morganbr 2018-06-03
If ```HashCode.Combine``` is only twice as slow as xor, it's probably worth doing since this xor would likely produce lots of collisions. I don't have a good sense of how often a Decimal is used as a Dictionary key though, which would dictate how important that is.