@jkotas 2018-06-03
I would work better to submit functional improvements / bug fixes in decimal to CoreCLR first, or at the same time as they are submitted to CoreRT. It will make testing and keeping track of what it fixed where easier.

Would it be possible to add this (together with what it depends on to make it compile) to `Decimal.DecCalc.cs` in CoreCLR?