@jkotas 2018-06-03
We do not randomize hashcodes of any built-in numeric types today. I would look at what to do about that separately from this issue. I do not think it makes sense to randomize for some and not for others.