@pentp 2018-06-04
I think I can refactor it as a self-contained function for CoreCLR without bringing all of `DecCalc` along.