@slavanap 2018-06-04
@pentp What input data do you use for performance tests? And how you run them?
For instance, for decimals ending with 23 zeros with scale = 23 there will be 5 invocations of Div96ByConst1XX functions, each of which contains 2 ulong divisions and 2 multiplications (1 for ulong and 1 for uint), that is more operations than in original CoreCLR implementation.