@pentp 2018-06-04
@slavanap I've been using the same approach as in #4452, which is somewhat biased toward normalized decimals. I'll try measuring with a different set of decimals from https://github.com/dotnet/corefx/pull/30116 and see if the perf is unacceptable.
All the divisions are by a constant so actually multiplications by the reciprocal.