@pentp 2018-06-04
With heavily denormalized decimals perf is following:
### x64
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 6.162 ns | 0.0159 ns | 0.0057 ns |   1.00 |
| CoreRT2 | 6.482 ns | 0.0128 ns | 0.0046 ns |   1.05 |
### x86
|  Method |      Mean |     Error |    StdDev | Scaled |
|-------- |----------:|----------:|----------:|-------:|
|  Native |  8.408 ns | 0.0426 ns | 0.0152 ns |   1.00 |
| CoreRT2 | 29.966 ns | 0.1981 ns | 0.0515 ns |   3.56 |

It looks like 32-bit RyuJIT doesn't use reciprocal multiplication for long divisions 😞