@jkotas 2018-06-04
> It looks like 32-bit RyuJIT doesn't use reciprocal multiplication for long divisions

Are you testing this on .NET Core 2.1?