@pentp 2018-06-04
Yes, latest coreclr source with debug JIT (for JitDump) with a release build:

```
if((low & 1) == 0 && iScale >= 1)
{
	const uint pow = 10;
	ulong div64 = high64 / pow;
	uint div = (uint)((((high64 - (uint)div64 * pow) << 32) | low) / pow);
	if(low == div * pow)
	{
		high64 = div64;
		low = div;
		iScale--;
	}
}
```
```
test     byte  ptr [ebp-18H], 1
jne      G_M65130_IG13
mov      dword ptr [ebp-14H], eax
cmp      eax, 1
jb       G_M65130_IG16
push     edi
push     esi
push     0
push     10
call     CORINFO_HELP_ULDIV
mov      dword ptr [ebp-80H], eax
mov      dword ptr [ebp-7CH], edx
mov      eax, dword ptr [ebp-80H]
mov      edx, dword ptr [ebp-7CH]
mov      dword ptr [ebp-48H], edx
mov      dword ptr [ebp-44H], eax
lea      ecx, [eax+4*eax]
shl      ecx, 1
mov      edx, esi
sub      edx, ecx
mov      ecx, dword ptr [ebp-18H]
mov      eax, ecx
or       eax, 0
or       edx, 0
push     edx
push     eax
push     0
push     10
call     CORINFO_HELP_ULDIV
mov      dword ptr [ebp-88H], eax
mov      dword ptr [ebp-84H], edx
mov      eax, dword ptr [ebp-88H]
mov      edx, dword ptr [ebp-84H]
lea      edx, [eax+4*eax]
shl      edx, 1
mov      ecx, dword ptr [ebp-18H]
cmp      ecx, edx
mov      dword ptr [ebp-18H], ecx
jne      SHORT G_M65130_IG15
mov      esi, dword ptr [ebp-44H]
mov      edi, dword ptr [ebp-48H]
mov      ecx, eax
mov      eax, dword ptr [ebp-14H]
dec      eax
mov      dword ptr [ebp-18H], ecx
```

There are some weird `xor eax, 0` lines also...