@jkotas 2018-06-04
Ah ok. I does not do it for 64-bit division. It does it for 32-bit division only. I am not sure whether the magic division would be profitable for 64-bit on x86. The C++ compiler does not use it in this case either.