@pentp 2018-06-05
I added a shortcut for 32-bit divisions that reduces `CORINFO_HELP_ULDIV` calls.
Refactored `VarDecDiv` to also use the new unscaling code and this brought x86 DIV from 1.03 to 0.98 on the benchmark.
With denormalized but still realistic numbers, x86 `GetHashCode` is 1.82x slower than currently - this feels like an acceptable trade-off considering that the current implementation is buggy.